drf_api/requirements.txt
Django>5.0,<5.2
djangorestframework==3.15.0,<3.15.2
django-cleanup==9.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg2-binary>=2.9,<3.0
PyJWT==2.9.0